+++ docs/release_notes.rst
@@ -11,7 +11,7 @@ Release notes
   - https://github.com/lucabaldini/aptapy/pull/53
   - https://github.com/lucabaldini/aptapy/pull/51
   - https://github.com/lucabaldini/aptapy/issues/50
-
+* Added `fit_iterative()` method to the `Gaussian` model class to perform iterative fitting.
 
 Version 0.14.0 (2025-11-14)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~