+++ docs/release_notes.rst
@@ -9,6 +9,7 @@ Release notes
 * Added the ``from_amptek_file()`` method to create a ``Histogram1d`` from an MCA8000A multichannel
   analyzer output file.
 * Pull requests merged and issues closed:
+
   - https://github.com/lucabaldini/aptapy/pull/54
   - https://github.com/lucabaldini/aptapy/pull/53
   - https://github.com/lucabaldini/aptapy/pull/51