+++ docs/release_notes.rst
@@ -2,6 +2,9 @@
 
 Release notes
 =============
+
+* Naming order for polynomial coefficients changed to match standard mathematical
+  notation (e.g., c2 for quadratic term, c1 for linear term, c0 for constant term).
 * Keeping track of the covariance matrix in the FitResult object.
 * Added numerical Jacobian calculation in the base class for fit models.
 * Added capability to plot confidence bands around fit models.
@@ -12,6 +15,7 @@ Release notes
   analyzer output file.
 * Pull requests merged and issues closed:
 
+  - https://github.com/lucabaldini/aptapy/pull/58
   - https://github.com/lucabaldini/aptapy/pull/56
   - https://github.com/lucabaldini/aptapy/pull/54
   - https://github.com/lucabaldini/aptapy/pull/53